# cmd/saas-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for WhatsApp Bot SaaS (Modular Architecture)",
        "title": "WhatsApp Bot SaaS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@whatsapp-saas.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/workflows": {
            "get": {
                "description": "Retrieve all workflows for a specific client",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "List workflows for a client",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new automation workflow for a client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Create a new workflow",
                "parameters": [
                    {
                        "description": "Workflow details",
                        "name": "workflow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.CreateWorkflowRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workflows/{id}": {
            "get": {
                "description": "Retrieve a specific workflow by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Get workflow by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing workflow",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Update a workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated workflow details",
                        "name": "workflow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.UpdateWorkflowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a workflow by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Delete a workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workflows/{id}/execute": {
            "post": {
                "description": "Trigger a workflow execution manually",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Manually execute a workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Trigger data",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.WorkflowExecutionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workflows/{id}/executions": {
            "get": {
                "description": "Retrieve execution history for a specific workflow",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Get workflow execution history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Limit number of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.Action": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "Action-specific configuration",
                    "type": "object",
                    "additionalProperties": true
                },
                "type": {
                    "description": "Action type: \"send_whatsapp\", \"update_database\", \"call_api\", etc.",
                    "type": "string"
                }
            }
        },
        "github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.Condition": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "Field to check (e.g., \"total_amount\", \"customer_type\")",
                    "type": "string"
                },
                "logic": {
                    "description": "\"AND\" or \"OR\" (default: \"AND\")",
                    "type": "string"
                },
                "operator": {
                    "description": "Operator: \"equals\", \"greater_than\", \"less_than\", \"contains\", etc.",
                    "type": "string"
                },
                "value": {
                    "description": "Value to compare against"
                }
            }
        },
        "github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.CreateWorkflowRequest": {
            "type": "object",
            "required": [
                "actions",
                "name",
                "trigger_config",
                "trigger_type"
            ],
            "properties": {
                "actions": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.Action"
                    }
                },
                "conditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.Condition"
                    }
                },
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "description": "Pointer to allow explicit false",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "trigger_config": {
                    "$ref": "#/definitions/github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.TriggerConfig"
                },
                "trigger_type": {
                    "type": "string",
                    "enum": [
                        "event",
                        "scheduled",
                        "manual"
                    ]
                }
            }
        },
        "github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.TriggerConfig": {
            "type": "object",
            "properties": {
                "event_name": {
                    "description": "For event triggers: \"transaction_created\", \"message_received\", etc.",
                    "type": "string"
                },
                "schedule": {
                    "description": "For scheduled triggers: cron expression \"0 18 * * *\"",
                    "type": "string"
                }
            }
        },
        "github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.UpdateWorkflowRequest": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.Action"
                    }
                },
                "conditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.Condition"
                    }
                },
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "trigger_config": {
                    "$ref": "#/definitions/github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.TriggerConfig"
                },
                "trigger_type": {
                    "type": "string",
                    "enum": [
                        "event",
                        "scheduled",
                        "manual"
                    ]
                }
            }
        },
        "github_com_MuhamadAgungGumelar_micro-system-ai-agent-be_internal_core_workflow.WorkflowExecutionRequest": {
            "type": "object",
            "properties": {
                "trigger_data": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
    }
}